Fix doulbe type

+++ cpp/include/nocturne_bits/geometry/geometry_utils.h
@@ -38,7 +38,7 @@ inline float AngleAdd(float lhs, float rhs) {
 }
 
 inline double AngleAdd(double lhs, double rhs) {
-  const float ret = std::fmod(lhs + rhs, kTau);
+  const double ret = std::fmod(lhs + rhs, kTau);
   return ret < 0.0 ? ret + kTau : ret;
 }
 
@@ -48,7 +48,7 @@ inline float AngleSub(float lhs, float rhs) {
 }
 
 inline double AngleSub(double lhs, double rhs) {
-  const float ret = std::fmod(lhs - rhs, kTau);
+  const double ret = std::fmod(lhs - rhs, kTau);
   return ret < 0.0 ? ret + kTau : ret;
 }
